Update README.md
## Known Issues

While we strive for a seamless experience, there are a few known issues in Beno House Finder App:

1. **Map Rendering Feature has not been implemented yet and will be implemented anytime soon:**
   - Occasionally, you won't be able to access the map functionality now.

3. **Android Version Compatibility:**
   - Some users on older Android versions may encounter minor UI glitches. We recommend using the latest Android OS for the best experience.

## Reporting Issues

If you encounter any issues not listed here or have suggestions for improvement, please [open an issue](link-to-issues). Your feedback is invaluable, and we are committed to enhancing your experience with Beno House Finder App.